pyproject.toml
[project]
name = "tester-agent"
version = "0.1.0"
description = "Proxima Tester Agent"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.52.0",
    "faker>=37.0.0",
    "python-dotenv>=1.1.0",
    "qwen-agent>=0.0.15",
]
